Feat : 필터링 데이터 패칭 구현 및, 서버 컴포넌트에서 호출하도록 분리
