+++ src/components/domains/perfumes/filter/index.tsx
@@ -1,73 +1,48 @@
 import React from "react";
 import { useFilterStore } from "@/lib/stores/useFilterStore";
-import Dropdown from "@/components/commons/dropdown/DropdownBase";
-
-// 필터 옵션 정의(Mock)
-export const FILTER_CONFIG = {
-  gender: {
-    label: "성별",
-    options: [
-      { label: "남성", value: "male" },
-      { label: "여성", value: "female" },
-    ],
-  },
-  brand: {
-    label: "브랜드",
-    options: [
-      {
-        label: "Dior",
-        value: "2c88f47d-47b5-4127-bcce-c54400b52dd4",
-      },
-      {
-        label: "Frederic Malle",
-        value: "608b9ece-7c4c-487b-bfe8-5646ec35fc98",
-      },
-    ],
-  },
-  notes: {
-    label: "노트",
-    options: [
-      { label: "Vanilla", value: "vanilla" },
-      { label: "Lemon", value: "Lemon" },
-    ],
-  },
-};
+import FilterDropdown from "./FilterDropdown";
+import { brands, perfume_notes } from "@prisma/client";
+import { JsonValue } from "@prisma/client/runtime/library";
 
 const typedKeys = <T extends Record<string, unknown>>(obj: T) =>
   Object.keys(obj) as (keyof T)[];
 
-const PurFumeFilter = () => {
-  const { filters, handleFilterChange, resetFilters } = useFilterStore();
+const PerFumeFilter = ({
+  brands,
+  notes,
+}: {
+  brands: brands[];
+  notes: perfume_notes[];
+}) => {
+  const { resetFilters } = useFilterStore();
 
-  const getSelectedOption = (category: keyof typeof FILTER_CONFIG) => {
-    const count = filters.get(category)?.size ?? 0;
-    const baseLabel = FILTER_CONFIG[category].label;
+  const toOption = (item: { id: string; name: JsonValue }) => ({
+    label: (item.name as { en: string }).en,
+    value: item.id,
+  });
 
-    return {
-      label: count > 0 ? `${baseLabel} ${count}` : baseLabel,
-      value: "",
-    };
+  const filterOptions = {
+    brand: brands.map(toOption),
+    notes: notes.map(toOption),
   };
+
+  const LABELS = {
+    brand: "브랜드",
+    notes: "노트",
+  };
+
   return (
     <>
       <div className="flex gap-5">
         {/* 필터 버튼 렌더링 */}
-        {typedKeys(FILTER_CONFIG).map((category) => {
-          const { options } = FILTER_CONFIG[category];
-
-          return (
-            <div key={category} className="flex flex-col items-start">
-              <Dropdown
-                selectedOption={getSelectedOption(category)}
-                options={options}
-                currentOption={Array.from(filters.get(category) ?? [])[0]}
-                handleChangeOption={(option) =>
-                  handleFilterChange(category, option.value)
-                }
-              />
-            </div>
-          );
-        })}
+        {typedKeys(filterOptions).map((category) => (
+          <FilterDropdown
+            key={category}
+            category={category}
+            label={LABELS[category]}
+            options={filterOptions[category]}
+          />
+        ))}
       </div>
       {/* 필터 목록 버튼 */}
       <button
@@ -81,4 +56,4 @@ const PurFumeFilter = () => {
   );
 };
 
-export default PurFumeFilter;
+export default PerFumeFilter;
